fix(zcash): small improvements
